+++ core/bh_fuse_price.cpp
@@ -24,7 +24,7 @@ If not, see <http://www.gnu.org/licenses/>.
 
 #include <bh_fuse_price.hpp>
 #include <bh_ir.hpp>
-#include <bh_pprint.h>
+#include <bh_pprint.hpp>
 
 using namespace std;
