+++ filter/noneremover/noneremover.cpp
@@ -19,7 +19,7 @@ If not, see <http://www.gnu.org/licenses/>.
 */
 
 #include <bh_component.h>
-#include <bh_pprint.h>
+#include <bh_pprint.hpp>
 
 using namespace std;
